Add ResourceLoader scripts, and an edit preview api
Add ResourceLoader scripts to the asset bundle, with basic modules, and allow
specific modules to load on pages (configurable).

Add a new edit preview endpoint to match the page view and include any allowed
resource loader modules in the output HTML.

Prepare a hook for the TabberTransclude extension, to transclude the contents
of the tabs instead of relying on the lazy loading of the tabs, to allow read
them if there's no internet connection.